requirements.txt
coverage==4.3.4
coveralls==1.1
Django==1.10.5
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
docopt==0.6.2
PyJWT==1.4.2
requests==2.12.5
